change entity interfaces to entity classes, add script
