+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/dao/UserRoleDAO.java
@@ -1,6 +1,5 @@
 package org.my.adventure.dao_manager.api.dao;
 
-import org.my.adventure.dao_manager.api.dao.CommonDAO;
 import org.my.adventure.dao_manager.api.entities.UserRole;
 
 /**
