+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/Common.java
@@ -1,9 +1,24 @@
 package org.my.adventure.dao_manager.api.entities;
 
+import javax.persistence.*;
+
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public interface Common {
-    Long getId();
-    void setId(Long id);
+@MappedSuperclass
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+public abstract class Common {
+    protected Long id;
+
+    @Id
+    @SequenceGenerator(name="sequence",sequenceName="entity_id_seq")
+    @GeneratedValue(strategy= GenerationType.SEQUENCE, generator="sequence")
+    @Column(name = "id", nullable = false, insertable = true, updatable = true)
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 }
