+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/Node.java
@@ -1,19 +1,65 @@
 package org.my.adventure.dao_manager.api.entities;
 
 import java.util.List;
+import javax.persistence.*;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface Node extends Common {
-    List<Transition> getTransitions();
-    void setTransitions(List<Transition> transitions);
-    String getName();
-    void setName(String name);
-    String getDescription();
-    void setDescription(String description);
-    Quest getQuestByQuestId();
-    void setQuestByQuestId(Quest questByQuestId);
-    List<Action> getActions();
-    void setActions(List<Action> actions);
+@Entity
+@Table(name = "node", schema = "public", catalog = "netcracker")
+public class Node extends Common{
+    private String name;
+    private String description;
+    private Quest questByQuestId;
+    private List<Action> Actions;
+    private List<Transition> transitions;
+
+    @OneToMany(mappedBy = "nodeByFromNode")
+    public List<Transition> getTransitions() {
+        return transitions;
+    }
+
+    public void setTransitions(List<Transition> transitions) {
+        this.transitions = transitions;
+    }
+
+    @Basic
+    @Column(name = "name", nullable = false, insertable = true, updatable = true, length = 200)
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Basic
+    @Column(name = "description", nullable = true, insertable = true, updatable = true, length = 2000)
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "quest_id", referencedColumnName = "id", nullable = false)
+    public Quest getQuestByQuestId() {
+        return questByQuestId;
+    }
+
+    public void setQuestByQuestId(Quest questByQuestId) {
+        this.questByQuestId = questByQuestId;
+    }
+
+    @OneToMany(mappedBy = "node")
+    public List<Action> getActions() {
+        return Actions;
+    }
+
+    public void setActions(List<Action> actions) {
+        Actions = actions;
+    }
 }
