+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/SavedItem.java
@@ -1,11 +1,32 @@
 package org.my.adventure.dao_manager.api.entities;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface SavedItem extends Common {
-    Game getGameByGameId();
-    void setGameByGameId(Game gameByGameId);
-    Item getItemByItemId();
-    void setItemByItemId(Item itemByItemId);
+import javax.persistence.*;
+@Entity
+@Table(name = "saved_item", schema = "public", catalog = "netcracker")
+public class SavedItem extends Common{
+    private Game gameByGameId;
+    private Item itemByItemId;
+
+    @ManyToOne
+    @JoinColumn(name = "game_id", referencedColumnName = "id", nullable = false)
+    public Game getGameByGameId() {
+        return gameByGameId;
+    }
+
+    public void setGameByGameId(Game gameByGameId) {
+        this.gameByGameId = gameByGameId;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "item_id", referencedColumnName = "id", nullable = false)
+    public Item getItemByItemId() {
+        return itemByItemId;
+    }
+
+    public void setItemByItemId(Item itemByItemId) {
+        this.itemByItemId = itemByItemId;
+    }
 }
