+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/SavedStat.java
@@ -1,13 +1,43 @@
 package org.my.adventure.dao_manager.api.entities;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface SavedStat extends Common {
-    Stat getStat();
-    void setStat(Stat stat);
-    float getValue();
-    void setValue(float value);
-    Game getGameByGameId();
-    void setGameByGameId(Game gameByGameId);
+import javax.persistence.*;
+@Entity
+@Table(name = "saved_stat", schema = "public", catalog = "netcracker")
+public class SavedStat extends Common{
+    private float value;
+    private Game gameByGameId;
+    private Stat stat;
+
+    @OneToOne
+    @JoinColumn(name = "stat_id", referencedColumnName = "id")
+    public Stat getStat() {
+        return stat;
+    }
+
+    public void setStat(Stat stat) {
+        this.stat = stat;
+    }
+
+    @Basic
+    @Column(name = "value", nullable = false, insertable = true, updatable = true, precision = 8)
+    public float getValue() {
+        return value;
+    }
+
+    public void setValue(float value) {
+        this.value = value;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "game_id", referencedColumnName = "id", nullable = false)
+    public Game getGameByGameId() {
+        return gameByGameId;
+    }
+
+    public void setGameByGameId(Game gameByGameId) {
+        this.gameByGameId = gameByGameId;
+    }
 }
