+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/Stat.java
@@ -1,15 +1,54 @@
 package org.my.adventure.dao_manager.api.entities;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface Stat extends Common {
-    String getName();
-    void setName(String name);
-    Float getValue();
-    void setValue(Float value);
-    String getDescription();
-    void setDescription(String description);
-    Quest getQuestByQuestId();
-    void setQuestByQuestId(Quest questByQuestId);
+import javax.persistence.*;
+@Entity
+@Table(name = "stat", schema = "public", catalog = "netcracker")
+public class Stat extends Common{
+    private String name;
+    private Float value;
+    private String description;
+    private Quest questByQuestId;
+
+    @Basic
+    @Column(name = "name", nullable = false, insertable = true, updatable = true, length = 200)
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Basic
+    @Column(name = "value", nullable = true, insertable = true, updatable = true, precision = 8)
+    public Float getValue() {
+        return value;
+    }
+
+    public void setValue(Float value) {
+        this.value = value;
+    }
+
+    @Basic
+    @Column(name = "description", nullable = true, insertable = true, updatable = true, length = 2000)
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "quest_id", referencedColumnName = "Id", nullable = false)
+    public Quest getQuestByQuestId() {
+        return questByQuestId;
+    }
+
+    public void setQuestByQuestId(Quest questByQuestId) {
+        this.questByQuestId = questByQuestId;
+    }
 }
