+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/Transition.java
@@ -1,13 +1,43 @@
 package org.my.adventure.dao_manager.api.entities;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface Transition extends Common {
-    String getCondition();
-    void setCondition(String condition);
-    Node getNodeByToNode();
-    void setNodeByToNode(Node nodeByToNode);
-    Node getNodeByFromNode();
-    void setNodeByFromNode(Node nodeByFromNode);
+import javax.persistence.*;
+@Entity
+@Table(name = "transition", schema = "public", catalog = "netcracker")
+public class Transition extends Common{
+    private String condition;
+    private Node nodeByToNode;
+    private Node nodeByFromNode;
+
+    @Basic
+    @Column(name = "condition", nullable = false, insertable = true, updatable = true, length = 1000)
+    public String getCondition() {
+        return condition;
+    }
+
+    public void setCondition(String condition) {
+        this.condition = condition;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "to_node", referencedColumnName = "id", nullable = false)
+    public Node getNodeByToNode() {
+        return nodeByToNode;
+    }
+
+    public void setNodeByToNode(Node nodeByToNode) {
+        this.nodeByToNode = nodeByToNode;
+    }
+
+    @ManyToOne
+    @JoinColumn(name = "from_node", referencedColumnName = "id", nullable = false)
+    public Node getNodeByFromNode() {
+        return nodeByFromNode;
+    }
+
+    public void setNodeByFromNode(Node nodeByFromNode) {
+        this.nodeByFromNode = nodeByFromNode;
+    }
 }
