+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/TypeOfResource.java
@@ -3,8 +3,19 @@ package org.my.adventure.dao_manager.api.entities;
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public interface TypeOfResource extends Common {
-    String getName();
+import javax.persistence.*;
+@Entity
+@Table(name = "type_of_resource", schema = "public", catalog = "netcracker")
+public class TypeOfResource extends Common{
+    private String name;
 
-    void setName(String name);
+    @Basic
+    @Column(name = "name", nullable = false, insertable = true, updatable = true, length = 100)
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 }
