+++ dao-manager/dao-manager-api/src/main/java/org/my/adventure/dao_manager/api/entities/UserRole.java
@@ -1,13 +1,32 @@
 package org.my.adventure.dao_manager.api.entities;
 
 import java.util.List;
+import javax.persistence.*;
 
 /**
- * Created by Дмитрий on 11.12.2015.
+ * Created by Дмитрий on 02.12.2015.
  */
-public interface UserRole extends Common{
-    List<User> getUsers();
-    void setUsers(List<User> users);
-    String getName();
-    void setName(String name);
+@Entity
+@Table(name = "user_role", schema = "public", catalog = "netcracker")
+public class UserRole extends Common {
+    private String name;
+    private List<User> users;
+
+    @OneToMany(mappedBy = "userRole")
+    public List<User> getUsers() {
+        return users;
+    }
+    public void setUsers(List<User> users) {
+        this.users = users;
+    }
+
+    @Basic
+    @Column(name = "name", nullable = false, insertable = true, updatable = true, length = 200)
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 }
