+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/ActionDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.ActionEntity;
+import org.my.adventure.dao_manager.api.entities.Action;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class ActionDAOImpl extends CommonDAOImpl<ActionEntity>{
+public class ActionDAOImpl extends CommonDAOImpl<Action>{
 }
