+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/GameDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.GameEntity;
+import org.my.adventure.dao_manager.api.entities.Game;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class GameDAOImpl extends CommonDAOImpl<GameEntity>{
+public class GameDAOImpl extends CommonDAOImpl<Game>{
 }
