+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/NodeDAOImpl.java
@@ -1,9 +1,10 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.NodeEntity;
+
+import org.my.adventure.dao_manager.api.entities.Node;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class NodeDAOImpl extends CommonDAOImpl<NodeEntity>{
+public class NodeDAOImpl extends CommonDAOImpl<Node>{
 }
