+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/QuestDAOImpl.java
@@ -1,9 +1,10 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.QuestEntity;
+import org.my.adventure.dao_manager.api.entities.Quest;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class QuestDAOImpl extends CommonDAOImpl<QuestEntity>{
+public class QuestDAOImpl extends CommonDAOImpl<Quest>{
+
 }
