+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/ResourceDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.ResourceEntity;
+import org.my.adventure.dao_manager.api.entities.Resource;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class ResourceDAOImpl extends CommonDAOImpl<ResourceEntity>{
+public class ResourceDAOImpl extends CommonDAOImpl<Resource>{
 }
