+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/SavedItemDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.SavedItemEntity;
+import org.my.adventure.dao_manager.api.entities.SavedItem;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class SavedItemDAOImpl extends CommonDAOImpl<SavedItemEntity>{
+public class SavedItemDAOImpl extends CommonDAOImpl<SavedItem>{
 }
