+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/SavedStatDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.SavedStatEntity;
+import org.my.adventure.dao_manager.api.entities.SavedStat;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class SavedStatDAOImpl extends CommonDAOImpl<SavedStatEntity>{
+public class SavedStatDAOImpl extends CommonDAOImpl<SavedStat>{
 }
