+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/StatDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.StatEntity;
+import org.my.adventure.dao_manager.api.entities.Stat;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class StatDAOImpl extends CommonDAOImpl<StatEntity> {
+public class StatDAOImpl extends CommonDAOImpl<Stat> {
 }
