+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/TransitionDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.TransitionEntity;
+import org.my.adventure.dao_manager.api.entities.Transition;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class TransitionDAOImpl extends CommonDAOImpl<TransitionEntity>{
+public class TransitionDAOImpl extends CommonDAOImpl<Transition>{
 }
