+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/TypeOfActionDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.TypeOfActionEntity;
+import org.my.adventure.dao_manager.api.entities.TypeOfAction;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class TypeOfActionDAOImpl extends CommonDAOImpl<TypeOfActionEntity>{
+public class TypeOfActionDAOImpl extends CommonDAOImpl<TypeOfAction>{
 }
