+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/TypeOfResourceDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.TypeOfResourceEntity;
+import org.my.adventure.dao_manager.api.entities.TypeOfResource;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class TypeOfResourceDAOImpl extends CommonDAOImpl<TypeOfResourceEntity>{
+public class TypeOfResourceDAOImpl extends CommonDAOImpl<TypeOfResource>{
 }
