+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/UserDAOImpl.java
@@ -1,9 +1,9 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.UserEntity;
+import org.my.adventure.dao_manager.api.entities.User;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class UserDAOImpl extends CommonDAOImpl<UserEntity>{
+public class UserDAOImpl extends CommonDAOImpl<User>{
 }
