+++ dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/dao/UserRoleDAOImpl.java
@@ -1,9 +1,10 @@
 package org.my.adventure.dao_manager.impl.dao;
 
-import org.my.adventure.dao_manager.impl.entities.UserRoleEntity;
+
+import org.my.adventure.dao_manager.api.entities.UserRole;
 
 /**
  * Created by Дмитрий on 11.12.2015.
  */
-public class UserRoleDAOImpl extends CommonDAOImpl<UserRoleEntity>{
+public class UserRoleDAOImpl extends CommonDAOImpl<UserRole>{
 }
