--- dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/entities/ItemEntity.java
@@ -1,45 +1,0 @@
-package org.my.adventure.dao_manager.impl.entities;
-
-import javax.persistence.*;
-
-/**
- * Created by Дмитрий on 02.12.2015.
- */
-@Entity
-@Table(name = "item", schema = "public", catalog = "netcracker")
-public class ItemEntity extends CommonEntity {
-    private String name;
-    private String description;
-    private QuestEntity quest;
-
-    @ManyToOne
-    @JoinColumn(name="quest_id")
-    public QuestEntity getQuest() {
-        return quest;
-    }
-
-    public void setQuest(QuestEntity quest) {
-        this.quest = quest;
-    }
-
-    @Basic
-    @Column(name = "name", nullable = false, insertable = true, updatable = true, length = 200)
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Basic
-    @Column(name = "description", nullable = true, insertable = true, updatable = true, length = 2000)
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-}
