--- dao-manager/dao-manager-impl/src/main/java/org/my/adventure/dao_manager/impl/entities/SavedItemEntity.java
@@ -1,33 +1,0 @@
-package org.my.adventure.dao_manager.impl.entities;
-
-import javax.persistence.*;
-
-/**
- * Created by Дмитрий on 02.12.2015.
- */
-@Entity
-@Table(name = "saved_item", schema = "public", catalog = "netcracker")
-public class SavedItemEntity extends CommonEntity{
-    private GameEntity gameByGameId;
-    private ItemEntity itemByItemId;
-
-    @ManyToOne
-    @JoinColumn(name = "game_id", referencedColumnName = "id", nullable = false)
-    public GameEntity getGameByGameId() {
-        return gameByGameId;
-    }
-
-    public void setGameByGameId(GameEntity gameByGameId) {
-        this.gameByGameId = gameByGameId;
-    }
-
-    @ManyToOne
-    @JoinColumn(name = "item_id", referencedColumnName = "id", nullable = false)
-    public ItemEntity getItemByItemId() {
-        return itemByItemId;
-    }
-
-    public void setItemByItemId(ItemEntity itemByItemId) {
-        this.itemByItemId = itemByItemId;
-    }
-}
