+++ dao-manager/dao-manager-impl/src/main/resources/hibernate.cfg.xml
@@ -11,19 +11,19 @@
       <property name="hibernate.hbm2ddl.auto">update</property>
       <property name="hibernate.show_sql">true</property>
       <property name="hibernate.dialect">org.hibernate.dialect.PostgreSQL9Dialect</property>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.ActionEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.GameEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.ItemEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.NodeEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.QuestEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.ResourceEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.SavedItemEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.SavedStatEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.StatEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.TransitionEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.UserEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.UserRoleEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.TypeOfActionEntity"/>
-      <mapping class="org.my.adventure.dao_manager.impl.entities.TypeOfResourceEntity"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Action"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Game"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Item"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Node"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Quest"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Resource"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.SavedItem"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.SavedStat"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Stat"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.Transition"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.User"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.UserRole"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.TypeOfAction"/>
+      <mapping class="org.my.adventure.dao_manager.api.entities.TypeOfResource"/>
   </session-factory>
 </hibernate-configuration>
