--- text-adventure-common/documents/Dmitriy Bogomolov/.~lock.Техническое задание Богомолов.docx#
@@ -1,1 +1,0 @@
-,al,al-home,26.11.2015 16:17,file:///home/al/.config/libreoffice/4;
